Fixed transaction handling problem.
The resetStatus method is called after the updates are complete. This method is used to indicate to the datastores that the all changes to the rows in it have been applied to the database. If we don’t call this method, future calls to update will cause the datastore to try to update the same rows again.